Created Google Storyboard and VC file so we can play with Storyboard
without having to commit to using it. If you like it, we can use it at the end of the week. We can work on this part of the storyboard without worrying about someone else being in the main storyboard and fucking it up. This is what Nathan walked through this morning.