Working on FriendVC desc. below
I am attempting to fix the double search bar issue, as well as no friends being displayed when the friendVC is pulled up. Since we have not added friends yet, I have hard-coded friends to display. I need to connect kids of the friends as well, which I haven't done, and am not entirely sure how to do.